requirements.txt
aiohttp==1.0.5
args==0.1.0
async-timeout==3.0.1
asyncio==3.4.3
awscli==1.16.116
boto3==1.9.106
botocore==1.12.106
certifi==2018.11.29
chardet==3.0.4
clint==0.5.1
colorama==0.3.9
configparser==3.7.3
discord.py==0.16.12
docutils==0.14
idna==2.8
jmespath==0.9.4
multidict==4.5.2
pyasn1==0.4.5
python-dateutil==2.8.0
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
websockets==3.4
